pyproject.toml
[project]
name = "abinet"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "datasets>=4.0.0",
    "datatrove>=0.6.0",
    #"flash-attn>=2.8.3",
    "flash-attn>=2.8.3",
    "huggingface_hub>=0.34.4",
    "hydra-core>=1.3.2",
    "ipykernel>=6.30.1",
    "jiwer>=4.0.0",
    "lightning>=2.6.0",
    "lmdb>=1.7.5",
    "matplotlib>=3.10.6",
    "opencv-python>=4.12.0.88",
    "pillow>=12.0.0",
    "schedulefree>=1.4.1",
    "timm>=1.0.22",
    "torch==2.8.0",
    "torchvision==0.23.0",
    "tqdm>=4.67.1",
    "transformers>=4.56.1",
    "wandb>=0.23.0",
]
